Importing one change that was made to revision 1.50.
git-svn-id: http://chianti.ucsd.edu/svn@2208 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/render/immed/src/cytoscape/render/immed/GraphGraphics.java
@@ -146,7 +146,7 @@ public final class GraphGraphics
   public final boolean contains(final byte shapeType,
                                 final float xMin, final float yMin,
                                 final float xMax, final float yMax,
-                                final double xQuery, final double yQuery)
+                                final float xQuery, final float yQuery)
   {
     if (m_debug) {
       if (!EventQueue.isDispatchThread())
